requirements.txt
Flask==2.2.2
requests==2.28.1
pymongo==4.3.3
python-dotenv==0.21.0